refactor(native): extract browser event records
